updated return values

+++ src/main/java/com/gb/lesson_9/controllers/TemperatureController.java
@@ -3,6 +3,7 @@ package com.gb.lesson_9.controllers;
 import com.gb.lesson_9.domain.TemperatureData;
 import com.gb.lesson_9.sevices.persistence.TemperatureRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -23,8 +24,10 @@ public class TemperatureController {
     }
 
     @GetMapping("/temperatures/{id}")
-    public TemperatureData getTemperatureById(@PathVariable int id) {
-        return temperatureRepository.getTemperatureById(id);
+    public ResponseEntity getTemperatureById(@PathVariable int id) {
+        TemperatureData result = temperatureRepository.getTemperatureById(id);
+        if (result != null) return ResponseEntity.ok(result);
+        return ResponseEntity.badRequest().body("Temperature was not found");
     }
 
     @PostMapping("/temperatures")
@@ -33,15 +36,37 @@ public class TemperatureController {
     }
 
     @PutMapping("/temperatures/{id}")
-    public void putTemperature(@RequestBody TemperatureData updatedTemperatureData, @PathVariable int id) {
-        TemperatureData temperatureData = temperatureRepository.getTemperatureById(id);
-        temperatureData.setTemperature(updatedTemperatureData.getTemperature());
-        temperatureData.setDate(updatedTemperatureData.getDate());
-        temperatureRepository.updateTemperature(temperatureData);
+    public ResponseEntity putTemperature(@RequestBody TemperatureData updatedTemperatureData, @PathVariable int id) {
+        TemperatureData result = temperatureRepository.getTemperatureById(id);
+        if (result != null) {
+            try {
+                result.setTemperature(updatedTemperatureData.getTemperature());
+                result.setDate(updatedTemperatureData.getDate());
+                result = temperatureRepository.updateTemperature(result);
+                return ResponseEntity.ok(result);
+            }
+            catch (Exception e) {
+                return ResponseEntity.badRequest().body(e);
+            }
+
+        }
+        return ResponseEntity.badRequest().body("Temperature was not found");
+
     }
 
     @DeleteMapping("/temperatures/{id}")
-    public void deleteTemperature(@PathVariable int id) {
-        temperatureRepository.deleteTemperatureById(id);
+    public ResponseEntity deleteTemperature(@PathVariable int id) {
+        TemperatureData result = temperatureRepository.getTemperatureById(id);
+        if (result != null) {
+            try {
+                temperatureRepository.deleteTemperatureById(id);
+                return ResponseEntity.ok(result);
+            }
+            catch (Exception e) {
+                return ResponseEntity.badRequest().body(e);
+            }
+        }
+        return ResponseEntity.badRequest().body("Temperature was not found");
+
     }
 }
